requirements.txt
supabase>=2.0.0
py-zk>=0.9.7
requests>=2.28.0
python-dotenv>=0.21.0
